Refactor PropType module imports
